# docs/swagger.yaml
definitions:
  browse.browseRequest:
    properties:
      favorite_only:
        type: boolean
      order:
        $ref: '#/definitions/meta.SortOrder'
      page:
        type: integer
      search:
        type: string
      sort:
        $ref: '#/definitions/meta.SortField'
      tag:
        type: string
    type: object
  browse.browseResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/browse.item'
        type: array
      pages:
        items:
          $ref: '#/definitions/browse.pageItem'
        type: array
      request:
        $ref: '#/definitions/browse.browseRequest'
      tag_favorite:
        type: boolean
    type: object
  browse.item:
    properties:
      create_time:
        type: string
      favorite:
        type: boolean
      id:
        type: integer
      is_read:
        type: boolean
      name:
        type: string
    type: object
  browse.pageItem:
    properties:
      content:
        type: string
      is_active:
        type: boolean
      is_enabled:
        type: boolean
      is_hidden_on_small:
        type: boolean
      link_url:
        type: string
    type: object
  browse.rescanLibraryResponse:
    properties:
      result:
        type: boolean
    type: object
  errors.Error:
    properties:
      cause: {}
      code:
        example: 0
        type: integer
      message:
        example: unknown error.
        type: string
    type: object
  meta.SortField:
    enum:
    - name
    - createTime
    type: string
    x-enum-varnames:
    - SortFieldName
    - SortFieldCreateTime
  meta.SortOrder:
    enum:
    - ascending
    - descending
    type: string
    x-enum-varnames:
    - SortOrderAscending
    - SortOrderDescending
  tag.listRequest:
    properties:
      favorite_only:
        type: boolean
    type: object
  tag.listResponse:
    properties:
      request:
        $ref: '#/definitions/tag.listRequest'
      tags:
        items:
          $ref: '#/definitions/tag.tagData'
        type: array
    type: object
  tag.setFavoriteRequest:
    properties:
      favorite:
        type: boolean
      tag:
        type: string
    type: object
  tag.setFavoriteResponse:
    properties:
      favorite:
        type: boolean
      request:
        $ref: '#/definitions/tag.setFavoriteRequest'
    type: object
  tag.tagData:
    properties:
      favorite:
        type: boolean
      id:
        type: integer
      name:
        type: string
    type: object
  view.setFavoriteRequest:
    properties:
      favorite:
        type: boolean
      name:
        type: string
    type: object
  view.setFavoriteResponse:
    properties:
      favorite:
        type: boolean
      request:
        $ref: '#/definitions/view.setFavoriteRequest'
    type: object
  view.updateCoverRequest:
    properties:
      index:
        type: integer
      name:
        type: string
    type: object
  view.updateCoverResponse:
    properties:
      request:
        $ref: '#/definitions/view.updateCoverRequest'
      success:
        type: boolean
    type: object
  view.viewRequest:
    properties:
      path:
        type: string
    type: object
  view.viewResponse:
    properties:
      browse_url:
        type: string
      favorite:
        type: boolean
      indices:
        items:
          type: integer
        type: array
      name:
        type: string
      request:
        $ref: '#/definitions/view.viewRequest'
      tags:
        items:
          type: string
        type: array
      version:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
info:
  contact: {}
  description: API Server for Mangaweb
  title: Mangaweb3 API
  version: "3.0"
paths:
  /browse:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        schema:
          $ref: '#/definitions/browse.browseRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/browse.browseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
  /browse/rescan_library:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/browse.rescanLibraryResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
  /browse/thumbnail:
    get:
      parameters:
      - description: name of the item
        in: query
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: body
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
  /tag/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tag.listRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.listResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
  /tag/set_favorite:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tag.setFavoriteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.setFavoriteResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
  /tag/thumbnail:
    get:
      parameters:
      - description: tag
        in: query
        name: tag
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: body
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
  /view:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/view.viewRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/view.viewResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
  /view/download:
    get:
      parameters:
      - description: name of the file
        in: query
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: body
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
  /view/get_image:
    get:
      parameters:
      - description: name of the item
        in: query
        name: name
        required: true
        type: string
      - description: width
        in: query
        name: width
        type: integer
      - description: height
        in: query
        name: height
        type: integer
      - description: index
        in: query
        name: i
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: body
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
  /view/set_favorite:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/view.setFavoriteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/view.setFavoriteResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
  /view/update_cover:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/view.updateCoverRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/view.updateCoverResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
swagger: "2.0"