# handler/browse/rescan_library.go
package browse

import (
	"net/http"

	"github.com/julienschmidt/httprouter"

	"github.com/rs/zerolog/log"
	"github.com/wutipong/mangaweb3-backend/handler"
	"github.com/wutipong/mangaweb3-backend/scheduler"
)

type rescanLibraryResponse struct {
	Result bool `json:"result"`
}

const (
	PathRescanLibrary = "/browse/rescan_library"
)

// @Success      200  {object}  browse.rescanLibraryResponse
// @Failure      500  {object}  errors.Error
// @Router /browse/rescan_library [get]
func RescanLibraryHandler(w http.ResponseWriter, r *http.Request, params httprouter.Params) {
	log.Info().Msg("Rescan library")

	scheduler.ScheduleScanLibrary()

	response := rescanLibraryResponse{
		Result: true,
	}

	handler.WriteResponse(w, response)
}